backend/requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
numpy==1.26.4
pandas==2.2.0
scipy==1.11.4
requests==2.31.0
tqdm==4.66.1
matplotlib==3.7.2
supabase==1.0.3
python-dotenv==1.0.0